removed unused import

+++ src/app/department/department.component.ts
@@ -190,6 +190,9 @@ export class DepartmentComponent implements OnInit {
    * @param departmentId 
    */
   removeDepartment(departmentId: number){
+   // document.getElementById('deleteConfirmModal').click();
+   var isConfirmed = window.confirm('Are you sure, you really want to delete this record?')
+   if(isConfirmed){
     this.departmentService.deleteDepartment(departmentId).subscribe(
       (response) => {
         console.log('exuected')
@@ -199,9 +202,14 @@ export class DepartmentComponent implements OnInit {
           setTimeout(()=>{
             window.location.reload();
           },1000)
+        }else{
+          this.toastr.error('error while deleting department with id: '+departmentId)
         }
       }
     )
+   }else{
+    this.toastr.warning('Department '+departmentId+' not deleted.')
+   }
   }
 
   /**
@@ -367,6 +375,27 @@ export class DepartmentComponent implements OnInit {
        this.updatedDepartmentLocationErrorInfo = '';
      }
      return this.isDepartmentLocationValid;
+   }
+
+   checkSubCheckBoxes(mainCheckBox: any){
+    //  var subCheckBoxesRow = document.querySelector('dataRow');
+    //  var subCheckBoxes = subCheckBoxesRow.querySelector("input[type='checkbox']") as HTMLInputElement;
+    //  subCheckBoxes.forEach( subCheckBox => {
+      
+    //  });
+    var departmentsToBeDeleted = [];
+   // var table = document.getElementById("myTable1")
+   // console.log(table)
+    //for(var i=0; i<tables.length; i++){
+    var rows = document.getElementsByTagName("tr");
+    //var subCheckBoxes = rows
+    for (var i = 0; i < rows.length; i++) {
+      var row = rows[i];
+      console.log("the value is" + rows[i]);
+      var subCheckbox = row.querySelector("input[type='checkbox']") as HTMLInputElement;
+      subCheckbox.checked = mainCheckBox.checked;
+      subCheckbox.click();
+    }
    }
 
    /**
